refactor: isolate Kotlin types by file
