+++ .agents/superpowers/plans/2026-07-11-proof-loss-p0-p3.md
