+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/hierarchy/TypeEdgeResolver.kt
@@ -3,12 +3,6 @@ package io.github.amichne.kast.shared.hierarchy
 import com.intellij.psi.PsiElement
 import io.github.amichne.kast.api.contract.Symbol
 
-/** An edge in the type hierarchy — links a PSI declaration to its resolved [symbol]. */
-data class TypeHierarchyEdge(
-    val target: PsiElement,
-    val symbol: Symbol,
-)
-
 /**
  * Backend-specific provider of supertype and subtype edges for a given PSI element.
  *
