+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/analysis/ProofLossAnalyzer.kt
@@ -3,41 +3,11 @@ package io.github.amichne.kast.shared.proofloss.analysis
 import io.github.amichne.kast.shared.proofloss.ir.Block
 import io.github.amichne.kast.shared.proofloss.ir.FunctionIr
 import io.github.amichne.kast.shared.proofloss.ir.PredicateCondition
-import io.github.amichne.kast.shared.proofloss.ir.FunctionId
-import io.github.amichne.kast.shared.proofloss.ir.SourceSpan
 import io.github.amichne.kast.shared.proofloss.ir.Statement
 import io.github.amichne.kast.shared.proofloss.ir.TrackedValueId
 import io.github.amichne.kast.shared.proofloss.ir.ValueExpression
-import io.github.amichne.kast.shared.proofloss.model.ArgumentIndex
-import io.github.amichne.kast.shared.proofloss.model.BoundaryId
 import io.github.amichne.kast.shared.proofloss.model.PredicateId
-import io.github.amichne.kast.shared.proofloss.model.CallableKey
 import io.github.amichne.kast.shared.proofloss.model.ProofModel
-
-sealed interface ProofEstablishment {
-    val location: SourceSpan
-
-    data class PredicateGuard(override val location: SourceSpan) : ProofEstablishment
-    data class MaterializationSuccess(
-        override val location: SourceSpan,
-        val producedValue: TrackedValueId,
-    ) : ProofEstablishment
-}
-
-data class ProofLossFinding(
-    val functionId: FunctionId,
-    val predicate: PredicateId,
-    val predicateCallable: CallableKey,
-    val boundary: BoundaryId,
-    val boundaryCallable: CallableKey,
-    val argumentIndex: ArgumentIndex,
-    val subject: TrackedValueId,
-    val boundaryArgument: TrackedValueId,
-    val proof: ProofEstablishment,
-    val boundaryCall: SourceSpan,
-    val valuePath: List<TrackedValueId>,
-    val suggestedMaterializers: Set<CallableKey>,
-)
 
 class ProofLossAnalyzer(private val model: ProofModel) {
     fun analyze(function: FunctionIr): List<ProofLossFinding> =
