+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/Block.kt
@@ -1,0 +1,5 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+data class Block(val statements: List<Statement>) {
+    constructor(vararg statements: Statement) : this(statements.toList())
+}
