+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/ExitKind.kt
@@ -1,0 +1,3 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+enum class ExitKind { RETURN, THROW }
