+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/FunctionIr.kt
@@ -1,0 +1,7 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+data class FunctionIr(
+    val id: FunctionId,
+    val parameters: Set<TrackedValueId>,
+    val body: Block,
+)
