+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/IrExtractor.kt
@@ -1,0 +1,5 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+fun interface IrExtractor<S> {
+    fun extract(source: S): ExtractionResult
+}
