+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/PredicateCondition.kt
@@ -1,0 +1,10 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+import io.github.amichne.kast.shared.proofloss.model.PredicateId
+
+data class PredicateCondition(
+    val predicate: PredicateId,
+    val subject: TrackedValueId,
+    val polarity: PredicatePolarity,
+    val location: SourceSpan,
+)
