+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/PredicatePolarity.kt
@@ -1,0 +1,11 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+enum class PredicatePolarity {
+    POSITIVE,
+    NEGATED;
+
+    fun predicateHoldsWhen(conditionResult: Boolean): Boolean = when (this) {
+        POSITIVE -> conditionResult
+        NEGATED -> !conditionResult
+    }
+}
