+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/SourceSpan.kt
@@ -1,0 +1,13 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+import io.github.amichne.kast.api.contract.NormalizedPath
+
+data class SourceSpan(
+    val filePath: NormalizedPath,
+    val startOffset: SourceOffset,
+    val endOffset: SourceOffset,
+) {
+    init {
+        require(endOffset >= startOffset)
+    }
+}
