+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/Statement.kt
@@ -2,36 +2,6 @@ package io.github.amichne.kast.shared.proofloss.ir
 
 import io.github.amichne.kast.shared.proofloss.model.ArgumentIndex
 import io.github.amichne.kast.shared.proofloss.model.BoundaryId
-import io.github.amichne.kast.shared.proofloss.model.PredicateId
-
-data class FunctionIr(
-    val id: FunctionId,
-    val parameters: Set<TrackedValueId>,
-    val body: Block,
-)
-
-data class Block(val statements: List<Statement>) {
-    constructor(vararg statements: Statement) : this(statements.toList())
-}
-
-data class PredicateCondition(
-    val predicate: PredicateId,
-    val subject: TrackedValueId,
-    val polarity: PredicatePolarity,
-    val location: SourceSpan,
-)
-
-enum class PredicatePolarity {
-    POSITIVE,
-    NEGATED;
-
-    fun predicateHoldsWhen(conditionResult: Boolean): Boolean = when (this) {
-        POSITIVE -> conditionResult
-        NEGATED -> !conditionResult
-    }
-}
-
-enum class ExitKind { RETURN, THROW }
 
 sealed interface Statement {
     val location: SourceSpan
