+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/ir/TrackedValueId.kt
@@ -1,0 +1,8 @@
+package io.github.amichne.kast.shared.proofloss.ir
+
+import io.github.amichne.kast.api.contract.NormalizedPath
+
+data class TrackedValueId(
+    val filePath: NormalizedPath,
+    val declarationOffset: SourceOffset,
+)
