+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/ArgumentIndexParseResult.kt
@@ -1,0 +1,6 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+sealed interface ArgumentIndexParseResult {
+    data class Valid(val value: ArgumentIndex) : ArgumentIndexParseResult
+    data class Negative(val value: Int) : ArgumentIndexParseResult
+}
