+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/BoundaryDescriptor.kt
@@ -1,0 +1,7 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+data class BoundaryDescriptor(
+    val id: BoundaryId,
+    val callable: CallableKey,
+    val obligations: List<Obligation>,
+)
