+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/CallableIdKey.kt
@@ -1,0 +1,8 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+@JvmInline
+value class CallableIdKey private constructor(val value: String) {
+    companion object {
+        fun parse(raw: String): TextParseResult<CallableIdKey> = parseText(raw, ::CallableIdKey)
+    }
+}
