+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/CallableKind.kt
@@ -1,0 +1,3 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+enum class CallableKind { FUNCTION }
