+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/CallableRole.kt
@@ -1,0 +1,3 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+enum class CallableRole { PREDICATE, MATERIALIZER, BOUNDARY }
