+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/MaterializerDescriptor.kt
@@ -1,0 +1,8 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+sealed interface MaterializerDescriptor {
+    val callable: CallableKey
+
+    data class Total(override val callable: CallableKey) : MaterializerDescriptor
+    data class NullableWithExit(override val callable: CallableKey) : MaterializerDescriptor
+}
