+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/ModelBuildResult.kt
@@ -1,0 +1,8 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+import io.github.amichne.kast.api.contract.NonEmptyList
+
+sealed interface ModelBuildResult {
+    data class Valid(val model: ProofModel) : ModelBuildResult
+    data class Invalid(val violations: NonEmptyList<ModelViolation>) : ModelBuildResult
+}
