+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/Obligation.kt
@@ -1,0 +1,6 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+data class Obligation(
+    val argumentIndex: ArgumentIndex,
+    val predicate: PredicateId,
+)
