+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/PredicateDescriptor.kt
@@ -1,0 +1,8 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+data class PredicateDescriptor(
+    val id: PredicateId,
+    val callable: CallableKey,
+    val subjectArgumentIndex: ArgumentIndex,
+    val materializers: Set<MaterializerDescriptor> = emptySet(),
+)
