+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/PredicateId.kt
@@ -1,0 +1,8 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+@JvmInline
+value class PredicateId private constructor(val value: String) {
+    companion object {
+        fun parse(raw: String): TextParseResult<PredicateId> = parseText(raw, ::PredicateId)
+    }
+}
