+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/ProofModel.kt
@@ -2,34 +2,6 @@ package io.github.amichne.kast.shared.proofloss.model
 
 import io.github.amichne.kast.api.contract.NonEmptyList
 
-sealed interface ModelViolation {
-    data class DuplicatePredicateId(val id: PredicateId) : ModelViolation
-    data class DuplicateBoundaryId(val id: BoundaryId) : ModelViolation
-    data class DuplicatePredicateCallable(val callable: CallableKey) : ModelViolation
-    data class DuplicateBoundaryCallable(val callable: CallableKey) : ModelViolation
-    data class DuplicateObligation(
-        val boundary: BoundaryId,
-        val obligation: Obligation,
-    ) : ModelViolation
-
-    data class UnknownPredicate(
-        val boundary: BoundaryId,
-        val predicate: PredicateId,
-    ) : ModelViolation
-
-    data class ConflictingCallableRoles(
-        val callable: CallableKey,
-        val roles: Set<CallableRole>,
-    ) : ModelViolation
-
-    data class ConflictingMaterializerPredicate(val callable: CallableKey) : ModelViolation
-}
-
-sealed interface ModelBuildResult {
-    data class Valid(val model: ProofModel) : ModelBuildResult
-    data class Invalid(val violations: NonEmptyList<ModelViolation>) : ModelBuildResult
-}
-
 class ProofModel private constructor(
     val predicates: List<PredicateDescriptor>,
     val boundaries: List<BoundaryDescriptor>,
