+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/proofloss/model/TextParseResult.kt
@@ -1,0 +1,15 @@
+package io.github.amichne.kast.shared.proofloss.model
+
+sealed interface TextParseResult<out T> {
+    data class Valid<T>(val value: T) : TextParseResult<T>
+    data object Blank : TextParseResult<Nothing>
+}
+
+internal inline fun <T> parseText(
+    raw: String,
+    construct: (String) -> T,
+): TextParseResult<T> =
+    raw.trim()
+        .takeIf(String::isNotEmpty)
+        ?.let { TextParseResult.Valid(construct(it)) }
+        ?: TextParseResult.Blank
