+++ backend-shared/src/test/kotlin/io/github/amichne/kast/shared/proofloss/model/ProofModelTest.kt
@@ -2,7 +2,6 @@ package io.github.amichne.kast.shared.proofloss.model
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertInstanceOf
-import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
